requirements.txt
streamlit
python-dotenv
youtube_transcript_api
pathlib
google-generativeai